Working on Approved Changes
Button in Header and Footer are a bit bigger

Headers are bigger and consistenet. Body text is smaller and a different font than header.

Body text in bio reformatted to look better and be easier to read

Unmute Text Button added to each thumbnail on hover, when muted. No functionality, yet.